requirements.txt
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.3.1
mongoengine==0.28.2
pymongo[srv]==4.7.2
python-decouple==3.8
PyJWT==2.8.0
bcrypt==4.1.3
Pillow==10.3.0
WeasyPrint==62.3
drf-spectacular==0.27.2
dnspython==2.6.1
gunicorn==22.0.0
whitenoise==6.7.0
